Cleaned entirely the branch so that only anything connected to my part is inside. All the files no more than 5 functions, and norm is left only for parser.c
