+++ include/scene.h
@@ -6,7 +6,7 @@
 /*   By: mevangel <mevangel@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/24 02:37:02 by fvoicu            #+#    #+#             */
-/*   Updated: 2024/06/16 16:50:53 by mevangel         ###   ########.fr       */
+/*   Updated: 2024/06/17 15:35:29 by mevangel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -45,12 +45,9 @@ typedef struct s_scene
 	t_camera	camera;
 	t_light		light;
 	t_ambilight	ambilight;
-	t_object	*objects; //dynamically allocated objects_nb times the t_object struct
+	t_object	*objects;
 	int			objects_nb;
 	t_viewport	viewport;
 }	t_scene;
-
-// t_scene	*init_scene(void);
-// void	destroy_scene(t_scene *scene);
 
 #endif
