+++ scenes/example.rt
@@ -1,11 +1,11 @@
 
-A	0.51				255,255,255
-C	-50.0,0,20			-1.0,1,0.99		18
+A	0.0001				255,255,255
+C	-50.0,0,20.			-1.0,1,1.0		180
 L	-40,0,30	1.0		10, 1, 20
+
 
 pl   0,0,0		0,1.0,0			255,0,225
 sp   0,0,20			20			255,0,0
 cy   50.0,0.0,20.6		0,0,1.0		14.2	21.42	10,1,255
-pl   0,10,-20		0,1.0,0			100,0,50
-cy   50.0,0.0,20.6   0.0,0.0,1.0   14.2   21.42   10,0,255
+
