+++ src/parsing/parsing_utils.c
@@ -1,19 +1,21 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   parse_utils.c                                      :+:      :+:    :+:   */
+/*   parsing_utils.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: mevangel <mevangel@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/06 14:35:19 by mevangel          #+#    #+#             */
-/*   Updated: 2024/06/16 18:04:00 by mevangel         ###   ########.fr       */
+/*   Updated: 2024/06/17 15:47:34 by mevangel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "miniRT.h"
 
 /**
- * @brief Counts and returns the amount of 'lines' the (passed as parameter) 2D array has.
+ * @brief Counts and returns the amount of 'lines' the 2D array has.
+ * @param array: the 2D array to be checked
+ * @returns: the number of strings this array contains
  */
 int	ft_2darray_size(char **array)
 {
@@ -26,13 +28,14 @@ int	ft_2darray_size(char **array)
 }
 
 /**
- * @brief Modifies the line passed as parameter, so that every comma or tab becomes space, 
- * so that after, the resulted string is splitted according to spaces. For example, 
- * the read line from the map: L	0,0,-5 	0.4		255,255,255
- * 					  becomes: L 0 0 -5  0.4  255 255 255
+ * @brief Modifies the line passed as parameter, so that every comma or tab
+ * becomes space, so that after, the resulted string is splitted according 
+ * to spaces. For example, the read line from the map:
+ * 			L	0,0,-5 	0.4		255,255,255
+ * becomes: L 0 0 -5  0.4  255 255 255
  * so that the numbers are properly splitted after, by one split call.
- * @param line: double pointer so that the string 'line' is passed by reference to be
- * modifiable.
+ * @param line: double pointer so that the string 'line' is passed by 
+ * reference to be modifiable.
  */
 void	modify_before_split(char **line)
 {
@@ -48,15 +51,15 @@ void	modify_before_split(char **line)
 }
 
 /**
- * @brief Checks if the string is a valid representation of a number, either integer or
- * float, rejecting strings like "0..234" or ".242" or "++123".
- * @param str: the string to be checked. Const cause it's only read here, not modified.
+ * @brief Checks if the string is a valid representation of a number, either
+ * integer or float, rejecting strings like "0..234" or ".242" or "++123".
+ * @param str: the string to be checked. Const cause it's not modified here.
  * @returns TRUE: if the string is a valid representation of a number.
  */
 static bool	ft_str_is_number(const char *str)
 {
-	bool found_dot;
-	
+	bool	found_dot;
+
 	if (!str)
 		return (false);
 	if (*str == '+' || *str == '-')
@@ -77,8 +80,8 @@ static bool	ft_str_is_number(const char *str)
 }
 
 /**
- * @brief Checks for an array of strings, if each one of the strings it contains, is a
- * valid representation of a number.
+ * @brief Checks for an array of strings, if each one of the strings it 
+ * contains, is a valid representation of a number.
  * @param array: the array of strings to be checked
  * @returns TRUE: if all the strings of the array are representing numbers.
  */
@@ -96,25 +99,26 @@ bool	array_has_only_numbers(char **array)
 }
 
 /**
- * @brief Exits the miniRT program, by doing the proper cleanup of two 2D arrays, one array of
- * t_object structs, and printing a relative error message to specify why it exited.
- * @param err_msg: the message that is printed to specify the error that occured
- * @param array1: the first 2D array to be freed
- * @param array2: the second 2D array to be freed
- * @param scene: pointer to access the carried 'objects' pointer, to be freed
- */
-void	ft_exit_miniRT(char *err_msg, char **array1, char **array2, t_scene *scene)
+*	@brief Checks the r, g, b values of the color and accepts only integer 
+*	values in the range [0-255]. Rejects negative numbers or floats and 
+*	protects from overflow in ft_atoi.
+*	@param color_arr: the array of the three strings for the r,g,b values.
+*	@returns true: if the the three r,g,b values do NOT meet the requirements
+*	for	valid color.
+*/
+bool	invalid_color(char **color_arr)
 {
-	if (err_msg)
+	int	i;
+	int	tmp;
+
+	i = -1;
+	while (++i < 3)
 	{
-		ft_putstr_fd(ERROR, 2);
-		ft_putendl_fd(err_msg, 2);
+		if (ft_strchr(color_arr[i], '.') || color_arr[i][0] == '-')
+			return (true);
+		tmp = ft_atoi(color_arr[i]);
+		if (tmp < 0 || tmp > 255)
+			return (true);
 	}
-	if (array1)
-		fv_free_array(array1);
-	if (array2)
-		fv_free_array(array2);
-	if (scene->objects)
-		free(scene->objects);
-	exit(EXIT_SUCCESS);
+	return (false);
 }
